+++ src/utils/routes.js
@@ -1,0 +1,2 @@
+export const transactionHistoryRoute = '/transactionHistory';
+export const exchangeRoute = '/exchange';